# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UserData:
    properties:
      full_name:
        type: string
      username:
        type: string
    type: object
  dto.UserRequest:
    properties:
      by:
        type: integer
      full_name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      username:
        type: string
    required:
    - full_name
    - password
    - role_id
    - username
    type: object
  dto.UserResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserData'
    type: object
  dto.UserSalaryRequest:
    properties:
      amount:
        type: number
      effective_from:
        type: string
      user_id:
        type: integer
    required:
    - amount
    - effective_from
    - user_id
    type: object
  dto.UserSalaryResponse:
    properties:
      amount:
        type: number
      full_name:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API untuk manajemen payslip.
  title: Payslip API
  version: "1.0"
paths:
...
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register user
      tags:
      - user
  /users/salary:
    post:
      consumes:
      - application/json
      description: Create a new employee salary
      parameters:
      - description: Employee salary payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserSalaryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserSalaryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: CreateSalary employee salary
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"